+++ python/ry/ryo3/_unindent.pyi
@@ -1,0 +1,6 @@
+"""ryo3-unindent types"""
+
+from __future__ import annotations
+
+def unindent(string: str) -> str: ...
+def unindent_bytes(string: bytes) -> bytes: ...
